+++ lib/page-list.js
@@ -8,7 +8,8 @@ const _ = require('lodash'),
   setup = require('./setup'),
   helpers = require('./services/elastic-helpers'),
   elastic = require('./services/elastic'),
-  utils = require('clay-utils'),
+  sites = require('./services/sites'),
+  utils = require('clayutils'),
   queue = require('./services/queue');
 
 var PAGES_INDEX;
@@ -60,33 +61,6 @@ function pageUriFromKey(key) {
 }
 
 /**
- * Find the proper site object using the sites service.
- *
- * @param  {String} key
- * @return {Object}
- */
-function findSite(key) {
-  // if (!_.get(setup, 'options.sites', null)) {
-  //   throw new Error('Womp');
-  // }
-
-  var site = setup.options.sites.getSiteFromPrefix(key),
-    parsedUrl;
-
-  // If the site is not found from the prefix we need to
-  // use the `host` value to find it.
-  if (!site) {
-    // Make sure it's prefixed with `//` so it can get parsed properly
-    // (Usually a problem when working locally)
-    key = _.includes(key, '//') ? key : `//${key}`;
-    parsedUrl = url.parse(key, true, true);
-    site = setup.options.sites.getSite(parsedUrl.host, '');
-  }
-
-  return site;
-}
-
-/**
  * Create the data for the page document
  * in the pages list
  *
@@ -96,19 +70,20 @@ function findSite(key) {
 function constructPageData(ops) {
   return _.map(ops, function (op) {
     var data = JSON.parse(op.value),
-      site = module.exports.findSite(op.key),
+      siteSlug = op.key.split('/').shift(),
       published = filters.isPublished(op),
       scheduled = filters.isScheduled(op),
       scheduledTime = scheduled ? new Date(parseInt(data.at, 10)) : null,
       publishTime = published ? new Date(parseInt(data.at, 10)) : null,
-      pageUri = pageUriFromKey(op.key);
+      pageUri = pageUriFromKey(op.key),
+      site = sites.getSiteBySlug(pageUri);
 
     return {
       type: op.type,
       key: pageUri,
       value: {
         createdAt: Date.now(),
-        uri: pageUri,
+        uri: utils.uriSlugToPrefix(pageUri, site),
         published: published,
         scheduled: scheduled,
         scheduledTime: scheduledTime,
@@ -117,7 +92,7 @@ function constructPageData(ops) {
         url: _.get(data, 'url', ''),
         title: '',
         authors: [],
-        siteSlug: site.slug
+        siteSlug
       }
     };
   });
@@ -290,11 +265,12 @@ function parseUpdateUrl(string) {
  * @return {Object}
  */
 function constructFindQuery(updateUrl) {
+  console.log('~~~~~~~~@@@@@@@@@', updateUrl);
   return {
     query: {
       multi_match: {
         query: parseUpdateUrl(updateUrl),
-        fields: ['url', '_id']
+        fields: ['url', 'uri']
       }
     }
   };
@@ -333,6 +309,7 @@ function findPageAndUpdate(reqObj, res) {
     res.status(200);
     res.json({ status: 'Success' });
   }).catch(function (err) {
+    console.log(err);
     // TODO: Improve this handling. Make more generic.
     res.status(400);
     res.json(err);
@@ -355,7 +332,6 @@ module.exports.updatePageEntry = updatePageEntry;
 module.exports.getPage = getPage;
 module.exports.pageExists = pageExists;
 module.exports.constructPageData = constructPageData;
-module.exports.findSite = findSite;
 module.exports.findPageAndUpdate = findPageAndUpdate;
 module.exports.updatePageListEntry = updatePageListEntry;
 module.exports.setPagesIndex = setPagesIndex;
