+++ lib/services/elastic-helpers.js
@@ -5,7 +5,9 @@ const _ = require('lodash'),
   bluebird = require('bluebird'),
   refProp = '_ref',
   log = require('./log').setup({file: __filename}),
-  setup = require('../setup');
+  setup = require('../setup'),
+  sites = require('./sites'),
+  { uriPrefixToSlug } = require('clayutils')
 
 /**
  * Used to test Elasticsearch data types
@@ -135,7 +137,10 @@ function resolveReferencesForPropertyOfStringType(propertyName) {
 
       if (value !== null && value !== undefined) {
         if (value[refProp]) {
-          result = setup.options.db.get(value[refProp]).then(JSON.parse).then(function (referencedValue) {
+          let site = sites(value[refProp]),
+            uri = site ? uriPrefixToSlug(value[refProp], site) : value[refProp];
+
+          result = setup.options.db.get(uri).then(JSON.parse).then(function (referencedValue) {
             op.value[propertyName] = convertObjectToString(referencedValue);
             return op;
           });
