+++ lib/services/filters.js
@@ -22,7 +22,7 @@ function filterRefs(op) {
  * @returns {boolean}
  */
 function isPageOp(op) {
-  return op.key.indexOf('/pages/') > 0;
+  return op.key.indexOf('/_pages/') > 0;
 }
 
 /**
