+++ lib/users-list.js
@@ -4,7 +4,7 @@ const _ = require('lodash'),
   elastic = require('./services/elastic'),
   helpers = require('./services/elastic-helpers'),
   log = require('./services/log').setup({file: __filename}),
-  USER_STRING = '/users/';
+  USER_STRING = '/_users/';
 
 var USERS_INDEX;
