+++ package.json
@@ -30,6 +30,7 @@
     "chalk": "^1.1.3",
     "clay-log": "^1.1.0",
     "clay-utils": "^1.1.1",
+    "clayutils": "^2.0.2",
     "elasticsearch": "13.0.0",
     "express": "^4.15.2",
     "glob": "^7.1.1",
